\name{automobile}
\alias{automobile}
\title{Data concerning automobile prices.}
\description{
  The \code{automobile} data frame has 201 rows and 25 columns and
  concerns automobiles in the 1985 Auto Imports Database. The response 
  variable, \code{price}, is the log selling price of the automobile. There
  are 7 categorical predictors and 17 continuous / integer predictors which
  are features of the automobiles. 41 automobiles have missing data in one
  or more of the feature entries. This dataset is true to the original except 
  with a few of the predictors dropped.
}
\usage{
data(automobile)
}
\source{
K Bache and M Lichman. UCI machine learning repository, 2013. 
http://archive.ics.uci.edu/ml/datasets/Automobile
}
\keyword{datasets}
