\name{benchmark_datasets}
\alias{ankara}
\alias{baseball}
\alias{boston}
\alias{compactiv}
\alias{ozone}
\alias{pole}
\alias{triazine}
\alias{wine.red}
\alias{wine.white}
\title{benchmark_datasets}
\description{
  Nine diverse datasets which were used for benchmarking bartMachine's out of sample performance in 
  the vignette for this package.
}
\usage{
data(benchmark_datasets)
}
\source{
See vignette for details.
}
\keyword{datasets}
