% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_package_builders.R
\name{build_extreme_bart_machine}
\alias{build_extreme_bart_machine}
\title{Builds an Extreme BART model}
\usage{
build_extreme_bart_machine(X = NULL, y = NULL, Xy = NULL,
  num_trees = 50, num_burn_in = 250, num_iterations_after_burn_in = 1000,
  hyper_a = 2, hyper_b = 1, hyper_q = 0.9, alpha = 0.95, beta = 2,
  mh_prob_steps = c(2.5, 2.5, 4)/9, debug_log = FALSE,
  run_in_sample = TRUE, cov_prior_vec = NULL,
  covariates_to_permute = NULL, use_missing_data = FALSE,
  use_missing_data_dummies_as_covars = FALSE,
  replace_missing_data_with_x_j_bar = FALSE,
  impute_missingness_with_rf_impute = FALSE, mem_cache_for_speed = TRUE,
  serialize = FALSE, seed = NULL, verbose = TRUE)
}
\arguments{
\item{X}{The matrix of covariates (n rows and p columns)}

\item{y}{The vector of extreme responses (length \code{n}). Make sure the minimum response is subtracted off 
prior to building the BART model.}

\item{Xy}{An alternative means to specify the covariates and response. This matrix has the response as the last column.}

\item{num_trees}{The number of trees in the BART model. Default is \code{50}.}

\item{num_burn_in}{How many initial Metropolis-within-Gibbs iterations are dropped. Default is \code{250}.}

\item{num_iterations_after_burn_in}{How many iterations to use. Default is \code{1000}.}

\item{hyper_a}{The hyperparameter \code{a} on the prior for lambda (inverse gamma). Default is \code{2}.}

\item{hyper_b}{The hyperparameter \code{b} on the prior for lambda (inverse gamma). Default is \code{1}.}

\item{hyper_q}{The probability the extreme BART model has a higher \code{k} value than a Weibull regression employing
the linear model of all \code{p} features in \code{X}. Recall that the \code{k} value in a Weibull distribution
is related to variance --- the higher the \code{k}, the lower the variance. Default is \code{0.9}.}

\item{alpha}{A hyperparemeter controlling tree depth via the probability of creating a new split.  Default is \code{0.95}.}

\item{beta}{A hyperparemeter controlling tree depth via the probability of creating a new split. It is the exponent
on the inverse depth of the tree at that point.  Default is \code{2}.}

\item{mh_prob_steps}{The hyperparameter vector controlling the probabilities of the tree modifications: \code{GROW}, \code{PRUNE} 
and \code{CHANGE}. The default is approximately \code{[0.28, 0.28, 0.44]}.}

\item{debug_log}{Print out a log file. Default is \code{FALSE}.}

\item{run_in_sample}{After model construction, should we compute the in-sample fits? Default is \code{TRUE}.}

\item{cov_prior_vec}{The hyperparameter vector controlling the probabilities of selecting the \code{p} covariates from \code{X}
during \code{GROW} steps. Default is \code{NULL} indicating uniform sampling.}

\item{covariates_to_permute}{This is a private parameter used by the testing functions only. Leave as \code{NULL} for expected behavior.}

\item{use_missing_data}{Should we use the "missing in attributes" algorithm for splitting naturally on missingness in \code{X}? 
Default is \code{TRUE}.}

\item{use_missing_data_dummies_as_covars}{Should we add dummy variables for missingness among the \code{p} features that feature missingness? Default is \code{TRUE}.}

\item{replace_missing_data_with_x_j_bar}{Should we add replace missingness among the \code{p} features that feature missingness with the average covariate value for
continuous predictors (and the modal value for categorical predictors)? Default is \code{FALSE}.}

\item{impute_missingness_with_rf_impute}{Should we add replace missingness among the \code{p} features that feature missingness with the predicted covariate value
arrived at using the \code{MissForest} algorithm? Default is \code{FALSE}.}

\item{mem_cache_for_speed}{Should we use memory caching for speed? This requires more memory and may not be possible with high \code{p} and 
\code{n}. Default is \code{TRUE}.}

\item{serialize}{Should we serialize the model to a file? Do so if you wish to save this model and transport it (load it up later).
Default is \code{FALSE} because it is slow.}

\item{seed}{The set seed value for reproducibility in the random algorithm. Default is \code{NULL} for no seed.}

\item{verbose}{Should we print out detailed messages? Default is \code{FALSE}.}
}
\value{
An object of type \code{extremeBartMachine}.
}
\description{
Builds an Extreme BART model
}
\author{
Kapelner
}
