% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremeBartMachine.R
\docType{package}
\name{extremeBartMachine}
\alias{extremeBartMachine}
\alias{extremeBartMachine-package}
\title{Extreme BART Machine}
\description{
Extreme BART Machine
}
\references{
TK
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}
}
\keyword{Learning,}
\keyword{Machine}
\keyword{Non-parametric}
\keyword{Tree-based}
\keyword{regression}
